requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
qrcode==7.4.2
reportlab==4.2.2
python-multipart==0.0.9
